+++ src/components/SingleProduct/CustomBreadcrumbs.tsx
@@ -1,45 +1,18 @@
 import { useLocation, useParams } from 'react-router-dom';
 import { Breadcrumb } from 'antd';
 import style from './Product.module.scss';
-import { AppRoute } from '@/enums/Route';
-import { laptopData, smartphoneData } from '../Card/constants';
+import { getBreadcrumbItems } from '@/components/helpers.ts';
+// import { AppRoute } from '@/enums/Route';
+// import { laptopData, smartphoneData } from '../Card/constants';
 
 export const CustomBreadcrumbs = () => {
   const { smartphone, id } = useParams();
-  const allProducts = [...smartphoneData, ...laptopData];
-  const currentProduct = allProducts.find((el) => id && el.id === +id);
   const location = useLocation();
 
-  let breadcrumbItems: object[];
-  if (location.pathname === '/basket') {
-    breadcrumbItems = [
-      {
-        title: 'Homepage',
-        href: `${AppRoute?.ROOT}`,
-      },
-      {
-        title: 'Basket',
-        href: '#',
-      },
-    ];
-  } else {
-    breadcrumbItems = [
-      {
-        title: 'Homepage',
-        href: `${AppRoute?.ROOT}`,
-      },
-      {
-        title: smartphone?.[0].toUpperCase().concat(smartphone?.slice(1)),
-        href: '#',
-      },
-      {
-        title: currentProduct?.title[0]
-          .toUpperCase()
-          .concat(currentProduct.title.slice(1)),
-        href: '#',
-      },
-    ];
-  }
+  const breadcrumbItems = getBreadcrumbItems(location.pathname, {
+    smartphone,
+    id,
+  });
 
   return (
     <Breadcrumb
